src/requirements.txt
alembic
Babel
docutils
Flask
Flask-Cors
Flask-Migrate
Flask-Moment
Flask-RESTful
Flask-Script
Flask-SQLAlchemy
Flask-WTF
gunicorn
Jinja2
psycopg2-binary
py
PyJWT
pylint
pyparsing
pytest
python-dateutil
python-dotenv
python-editor
python-jose
PyYAML
rsa
SQLAlchemy
urllib3
Werkzeug

